config : ProfileInput Text Control들 폴더링 변경
